refactor: handler-mock 데이터 분리
