+++ src/apis/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useAuthorResponsesQuery'
 export * from './useRecipientResponsesQuery'
 export * from './useSingleAuthorResponse'
+export * from './useSingleRecipientResponse'
