+++ src/mocks/handlers/index.ts
@@ -1,1 +1,2 @@
 export * from './test'
+export { default as CreatedReviewManageHandler } from './CreatedReviewManage'
