+++ src/mocks/handlers/test.ts
@@ -1,17 +1,3 @@
-import { rest } from 'msw'
+import createdReviewManagehandlers from '@/mocks/handlers/CreatedReviewManage'
 
-const todos = ['먹기', '자기', '놀기']
-
-export const handlers = [
-  // 할일 목록
-  rest.get('/todos', (_, res, ctx) => {
-    return res(ctx.status(200), ctx.json(todos))
-  }),
-
-  // 할일 추가
-  rest.post('/todos', async (req, res, ctx) => {
-    todos.push(await req.json())
-
-    return res(ctx.status(201))
-  }),
-]
+export const handlers = [...createdReviewManagehandlers]
